Update coub.sh
-) added --sort option
Tags and channels can now be downloaded in a special order. This is mainly useful in combination with --limit-num (e.g. download the 50 most viewed videos with a certain tag).
-) quick fix for a bug from the previous commit
A small mistake, but very embarrassing. During channel downloads, all normal coubs would result in a null link. With this fix the average speedup for channel downloads was reduced to ~150%.