Update coub.sh
-) another fix for the new timeline parsing (no more null links at the end of channel downloads)
-) improved check for already downloaded coubs
The old check looked if 
a) the coub link is present in the used archive file or 
b) a coub with the same name (%id%) can be found in the destination directory
Since the custom output formatting commit, this check was done AFTER downloading the coub's information json (via curl) in order to check against the final output name. This prevented rapid skipping. Now the script does the old check (existence in archive or existence of %id% filename, in case of default output formatting) before downloading the coub's information json. In case of custom output formatting an additional check is done afterwards, which is still quite slow. So for custom output formatting, the usage of an archive file is strongly recommended.