Add files via upload
coub.sh:
-) removed check for the preview command
Using anything but the vanilla command (i.e. no additional options) would throw an error as well. It also never protected the user from using non media player commands as long as they could be executed.
-) archive files now use the coub id instead of the provided link (fixes possible duplicates)
-) unexpected missing files in the preview function won't cause a hard exit now
-) --sleep is now ignored for the first coub
-) other small changes and typo fixes

coub.py:
New port of the original script to Python. It's planned as a 1:1 port as far as the functionality and inner workings go, but there are a few exceptions. Mainly the lack of workarounds that aren't necessary in Python and the replacement of 3rd party tools with Python modules.
Advanced input methods (lists, channels, tags), their related options and custom output formatting aren't implemented yet, but will soon follow.
This version isn't supposed to replace the original script, but to provide an (easily accessible) alternative for Mac and Windows.