Adept to Coub's database/API upgrade
coub.py:

Added fallback for --mobile. The old base64 URL gets rarer by the minute, so the script falls back to the html5 streams, if the new direct URL is detected (there's no need to download the mobile version, as it's now the same file as html5 med). This means it might be wise to use --worstvideo in addition to mobile, if you expect ~360p output.
Otherwise not much changed. This version is only here for compatibility reasons, because not all html5 videos are yet stored in a non-broken state. Especially html5 med is somewhat likely to need fixing. Already correct streams still get the additional bytes added at the beginning, but that shouldn't cause any issues (at least it didn't, when I played the downloaded coubs with both mpv and VLC).

coub_v2.py:

The new script. It provides the same functionality as coub.py, with a few exceptions.
-) no video repair
More and more video streams are stored correctly, so there's no further need to fix them.
-) no --mobile option
mobile video and html5 med are the same file (previously the mobile version was of slightly higher quality and didn't have the watermark), so there's no need to bother with mobile video for now. It might get reintroduced, once mobile AAC audio gets more common.

There are also various little to moderate code changes.
-) Options class only holds options (no more functions)
-) Input data gets stored in an new class (also handles parsing)
-) f-string for help text
-) cleaned up the command line parsing a bit
-) less overly-verbose variable names
-) only execute main function, if script gets run directly (not sure why anybody would do anything but that, but apparently it's good practice)