Delete coub.sh
It's been bothering me for a while that coub.sh is still present, when it offers almost no benefits over coub.py. Only --limit-rate kept it alive until now. However 3 scripts with basically the same functionality is just insane. Also I can't be bothered to port all new changes over to Bash.

It was fun writing it, but now the prototype has to go.