Add AAC audio support
Coub now serves AAC audio for mobile devices (in addition to MP3). This commit adds said audio version to the list of possible audio streams as well as two command line options to further tune the probability of downloading AAC audio. This is all described in more detail in the README.