Add support for 'share' video
The --share option allows to download the special video version labelled as 'share' (comes with both video and audio already muxed together). Quality-wise it lies right in the middle between the best and worst available video/audio. In addition it also cuts the audio short, which is sometimes useful, but often not. I don't see much reason to download these videos, but for completeness sake here's the option to do it.

Also fixes a bug (oversight in the last commit), when no video or audio streams were detected (i.e. the coub is unavailable or has no audio).