Simplify sort order handling
I made two very wrong assumptions when it comes to sort orders.

1.) Invalid sort values should be avoided.
In reality the Coub API simply ignores invalid values. There's no need for extensive checking, especially when it comes to different timelines supporting different sort values. So get rid of all the unnecessary clutter in CoubInputData.parse_timeline(). There's an additional check in check_options() which will be kept, as it can help to avoid typos.
2.) "newest" is the default sort order.
Not sure how I came up with this nonsense. Nothing besides coub searches even supports the "newest" sort order. The default sort order is probably "newest_popular", but in the end there's no need to provide a default value for the script. So set the sort order by default to None.

Since I was already working on the sort order option, I also introduced a new help text ('--sort help') to slim down the normal help. Some of the mentioned input methods are not yet available, but they will be in the near future.