Add hot section download
The user can now download coubs from Coub's hot section (currently most trending/popular) with the --hot option. Similar to tags, the hot section API endpoint restricts the page number to 99. Everything afterwards will redirect to page 1, which limits the max. number of downloadable coubs to 2475.

On a side node I removed the page_limit option. The initial thought was to allow advanced users to limit the max. number of parsed pages per timeline input, but in the end it was only used for tags (and now the hot section) anyway.