Asynchronous I/O
Coubs are now processed in batches. All coubs within a batch run through
a collective preproccesing, download and post-processing stage,
effectively separating the previous processing chain into individual
blocks.
This is especially useful as the actual download of the media streams
isn't buried deep within the processing chain anymore. Initially the
goal was to fully utilize a user's network connection by making the
download multithreaded, but excessive RAM usage (several hundred MB),
system stalling and an overall worse performance were the result. So
my focus shifted from multithreading to asynchronous I/O.

This is all fairly new to me and there's a lot of room for
improvements, but it works well so far. Still, using anything but the
master branch is discouraged.