urllib as fallback for aiohttp
It was always my goal to provide a script that works with a default
Python environment, so that people, who don't want to clutter their
system with otherwise unneeded packages or are not familiar with
Python package management, can still use it. Relying on aiohttp
exclusively went against this goal.
This commit reintroduces the old download method of sequential stream
downloads with urllib as fallback for aiohttp.