Check for stream corruption
Downloading several files at once instead of one by one has one major
flaw: Problems with the connection aren't isolated to a single file.
With the old version there was little need to check for a stream's
integrity. A dropped connection or user interrupt could at worst corrupt
a single coub and if Coub were to delete an already parsed link before
the script managed to download it, then it would just count as
unavailable.
The situation is different now. Interrupting the download process can
affect many coubs. Having a longer continuous download also increases
the chances of something happening to the connection. In particular I
had problems with Coub deleting coubs, while I was still downloading
them, which led to corrupted streams because of missing chunks (and
ugly FFmpeg errors in the terminal output).
Therefore it's time to introduce a check for stream corruption. This is
really primitive. Just run the 1st second of a stream through FFmpeg. If
it throws a typical error, discard the coub as corrupted