Timeline parsing tweaks
This is the first step to get away from the "do first, ask questions
later" approach, that involves lots and lots of error handling. Some-
times it's easier, but there's no need for endless try-except blocks,
just because I'm too lazy to test for the existence of a key/item
before assignment/comparison.