Combine async/sync code
Until now each async function had a separate sync counterpart, which led
to a lot of duplicate code, since the only real difference was the usage
of urllib/aiohttp (and to a lesser extent open()/aiofiles).
Now the async functions used by the coub buffer have the fallback to
urllib/open() included and are run regardless of the existence of
aiohttp/aiofiles. This also means that asyncio is now mandatory, but it
is included in the standard library since Python 3.4, so that shouldn't
be a problem. The only real difference for the end user is that the
required min. Python version was increased to 3.7 (which was previously
only optional).

The async parsing code will soon follow suit.