Clean up async timeline parsing
Timeline parsing became a huge and ugly mess after I made it async.
This should make things a lot better. As with the previous commit, async
functions now also contain the fallback in case of aiohttp not being
available.
Additionally the template for timeline page requests was moved to its
own function and the max. request limit (when --limit-num is being used)
now gets calculated properly.