Make coub processing asynchronous task
Using a buffer to bundle all IO-bound tasks was a workaround because of
my lacking experience (i.e. none) with asynchronous IO. While it worked
well for all networking tasks, it also created a new bottleneck:
The muxing of the separate audio/video streams.
Merging the separate streams requires only a little amount of resources,
but because you potentially output hundreds MB worth of data, you're
still IO-bound because of your disk's writing speed. So the merging step
should ideally be done as soon as possible, while other coubs are still
downloading.

This commit makes the entire process chain for a single coub (existence
check, API request, API response parsing, download, integrity check,
muxing, archival and preview) a coroutine. While most steps are not
awaitable, it allows the script to parse/check/merge/... one coub, while
others are still downloading. Other advantages are that a single aiohttp
ClientSession covers the entire coub processing stage and that we can
get rid of the buffer class once and for all.

These changes are barely tested and the terminal output is still
bare-bones, but it works on a basic level.