Improve error handling
The most important scenarios to handle properly are
  a) user interrupts
  b) no network connection
  c) losing the network connection at some point

This commit focuses on a) and b).

The new exception rules in Coub.parse() and save_stream() lead to an
immediate exit after use interrupts. No more additional "unavailable"
error messages.
On the other hand, trying to parse/download coubs without a working
connection will mark all coubs as unavailable (just like with old
coub.py versions).

Timeline parsing checks if a given timeline URL is valid, but also if
an API request can be made. If not then the script assumes no working
connection and quits.

The two big problems that are left are the annoying "cancelled future"
error messages when interrupting the script while using urllib and
losing the connection during the download process in general, which
stalls the script indefinitely.