Introduce retries after lost connection
Coub is pretty gallant when it comes to the number of coubs you
download, but they get iffy when you start working with several
connections.
Until now it meant script failure if Coub would disconnect you, which
is rather annoying for larger downloads that you keep in the background,
while you do something else.

Now the script tries to run the processing function again, until it
finishes successfully or a retry limit is reached (which can be also 0
or infinite). Retries are done via recursion, so in theory the script is
limited to 1000 retries (unless you overwrite Python's default recursion
limit), but that's beyond anything sensible anyway.
Important to note is that all of this hinges on the usage of aiohttp,
which has several reasons. Most importantly I haven't seen it happen
once that Coub disconnected me while using urllib. Additionally I simply
don't know what reason (of the URLError object) to check against to
catch such an occurrence.

This commit also simplifies some of aiohttp's exception handling. Until
now I was under the wrong impression to catch unavailable coubs. In
reality there's no exception when aiohttp encounters an unavailable site
(unlike urllib). You'd have to raise it from the returned response
object and even then it wouldn't be the exception type I checked against
until now.
In the end all of it is superfluous, because stream_lists() will catch
unavailable coubs during parsing and valid_stream() after the download
attempt.