Add connection check
Terminate the script early on or before starting a new retry, if it can
not connect to coub.com. This prevents a potential infinite recursions
(i.e. until Python's recursion limit is reached) when no network
connection is present or Coub is simply down. But even for finite retry
attempts it's more pleasant to receive a single error message than the
script spamming "Retrying..." warnings.

As far as the actual check is concerned, I decided to just use urllib's
urlopen. From what I understand there are more efficient ways to do
this, but why bother with low level networking stuff just to save the
fraction of a second execution time? This check only gets called a
handful of times anyway.