Check connection at a later time
The connection check was made even before the command line arguments got
parsed. Not much a problem during normal usage, but if you called the
script without arguments to get the help text (the lazy alternative to
using -h/--help) then there was a noticeable (and quite annoying) delay.

Now the connection check is made right before the parsing process, which
is the first step that requires a working connection in case of timeline
parsing.