Remove --sort option
The way Coub handles sorting is a mess. Sometimes sorting is handled by
the order_by parameter, sometimes by calling a completely different
endpoint. Sometimes the sorting method is indicated by a different URL,
sometimes not. Sometimes the default sorting method is a specific
sorting method, sometimes the default is actually the lack of any
sorting method. And to make matters even worse, most sorting methods
only apply to one or two input sources and get ignored by the rest.

Therefore get rid of the global --sort option. The goal is to eventually
include the sorting method on a per-input basis.