Try to categorize unspecified input
Everything but direct coub links required the usage of options to
specify the link type. Now all non-options (no leading '-') will be
treated as input. The user_input object is responsible to map these
unspecified links to the right type.
The mapping process is still far from being perfect, because many valid
links (such as the start page, communities, etc.) are unsupported and
get treated as channels (since channel URLs have no unique identifiers).

The input options (-t/--tags, etc.) will remain and allow the user to
bypass the mapping process. This is useful if one doesn't want to paste
a valid URL to e.g. search for a specific tag or keyword.