Replace categories with communities
Another misunderstanding due to the outdated API documentation. Most
people will know Coub's communities (Animals & Pets, Anime, etc.). I was
under the wrong impression that Coub used the explore timeline to list
coubs of a specific community, but they have an undocumented community
timeline for that. The explore timeline also supports communities, it
may even provide the same results as the community timeline, but I
rather try to emulate coub.com (incl. vocabulary), because that's what
people know.

Another reason for this change is that I can't trust the documentation
when it comes to supported sort orders, while I can be sure to be on the
right track when copying coub.com.
However, sorting is a challenge for another day. For now we stick with
newest first (https://coub.com/community/<community>/fresh).