Reintroduce sorting for channel downloads
This is the first glimpse of an overhauled sorting system. Channels are
the first timeline input type to get support, but the other types will
soon follow, which is why I don't refer to channels exclusively in the
next section.

The biggest change for the user is that there's not a global option to
handle sorting. Instead each input can now be sorted on its own, by
adding the sort order with a leading # after the input link/keyword.
For example

  -) https://coub.com/example-channel#only-recoubs
  -) -c example-channel#no-recoubs

If no sort order is specified, the script will use a default sort order
as specified in the Options object (advanced setting). Invalid sort
orders on the other hand are treated as errors and will label the
timeline as invalid. This might sound extreme compared to using the
default sort order as fallback, but depending on the input type,
different sort orders can not only alter the succession in which coubs
get parsed, but also what coubs will be listed (e.g. the most popular
category of a community may return 300-5000 links depending on the used
timescale).

Internally the process of assembling the API request will be split among
methods in ParseableTimeline, since Coub's formatting approach is far
from being consistent. Each input type will have its own method. The
keywords to label the different sorting orders will be kept close to
Coub's naming scheme, but I'll make exceptions if necessary. For example
using "singles" to not download a channel's recoubs is confusing,
because no normal user will know that the API keeps track of a coub's
type with Coub::Simple and Coub::Recoub.