Make sort order keywords independent
Tying the keyword to indicate a sort order to Coub's internal name for
said sort order was probably not the smartest idea. People know the
phrasing used on coub.com and not what API calls are made under the
hood.
By using dictionaries instead of lists to store all supported sort
orders, the keywords are independent of the internal phrasing and are
now closer to what people are used to on coub.com.