Prevent connection attempt for invalid timelines
The timeline template method may mark a timeline as invalid (i.e. an
unknown sort order was requested), but that only prevented the actual
parsing process and not the request for the overall page count.
Depending on the timeline type this could lead to unnecessary warnings
because of an incomplete request template (a working template isn't
guaranteed if an unsupported sorting/selection method was requested),
but in all cases it wastes time by making an unnecessary connection
attempt.