Fix wrong behavior for exotic tags/searches
Exotic might be a bit of a stretch here. This commit gets rid of 2 basic
assumptions that simply aren't true.

  1) There are no slashes in an input identifier (e.g tag name).

Since normalize_link() provides us with a specific URL structure, we can
remove the base part of a search/tag/community/etc. URL instead of just
using the string after the last slash.

  2) Using urllib's quoting function on copied links is a good idea.

Using the quoting function repeatedly on the same string leads to
garbled nonsense if an escape sequence (%xx) is already present, because
the percent sign gets replaced with "%20" each time. Using the reverse
unquote function beforehand solves this problem.