Stop option parsing after --
Not sure how useful this really is. In theory there might be a channel
out there somewhere that starts with a dash, which would be treated as
an option, if the name is provided by itself (which should otherwise
create a channel URL). Of course using the full URL or -c NAME would
also work, so this is a non-issue.

The real reason for this commit is that somebody recently told me that
I should stop parsing the command line manually and just switch to
argparse, because manual parsing misses the small subtleties that make
a good CLI tool, like -- to stop parsing options.
Well, perhaps the script is missing some of those small subtleties, but
at least his specific example can be implemented easily enough. So let's
make a passive-aggressive commit for Christmas.