Avoid parsing of duplicate timelines
The script removes duplicates from the final list of parsed links, but
there was no check if a timeline was already present in the timeline
list. If one were to accidentally specify a timeline twice, then it was
also parsed twice. Imagine this happens for a search with a general word
like "music" which provides over 50.000 links.

Important to note is that a timeline is only skipped if type, id and
sort order all match with an already existing timeline. Every other
combination can lead to different link lists.