Read archive file only once
For each coub the whole archive file was read. Then the script looped
over every line of the archive's content, until a match was found or EOF
was reached. This doesn't just sound inefficient, it really was.
Especially if one didn't just use archives for a handful of downloads,
but e.g. to follow the uploads to a community, which could quickly amass
archives with tens of thousands of logged IDs.

The inefficient match check was simply an oversight. Reading the archive
for every coub on the other hand was a design choice, from back when
duplicates weren't removed from the parsed link list. Reading from and
writing to the archive for each coub ensured that duplicate coubs within
a script instance would be quickly skipped as already existing. That is
not necessary anymore.
Instead read the archive file once in the beginning and keep its content
stored. Now we only need to check against this fixed list instead of
accessing the archive hundreds (if not thousands) of times.