Add option to change container format
Using MKV as standard container for the final remux (audio and video
combined) was supposed to make a clear distinction, between the
downloaded video stream (MP4), audio stream (MP3/M4A) and the final
output extension-wise. In other words I was too lazy to handle the
scenario in which the downloaded video stream and the final output would
have the exact same name.

Now the extension of the merged file can be changed in the advanced
settings section. The only requirement is that the container is able to
hold AVC video and MP3/AAC audio.

Fixes: #6