Request link quantity from timeline
To ensure adherence to the link limit (set via --limit-num) each parsing
related method constantly checked against the limit to stop adding links
as soon as the parsed link list reached its max. size.
This required the size of the parsed link list to be available for all
parsing methods, which is another reason, why all parsing is done in a
separate object.

Instead of checking whether or not the limit was reached, request only a
limited number of links in the first place. This is done in preparation
for the next commit.