Overwrite output link lists
A while ago I changed the default behavior for outputting link lists
(--write-list) from overwriting the target file to appending to it. Back
then it seemed like the better choice (being able to build lists over
time), but right now I find it rather annoying. Especially since it
makes it impossible to weed out duplicates in an existing list, without
writing the links to a different file. So let's revert the default to
overwriting output files.

In the end I can see both ways as being practical in some situations, so
I do what I always do, when I can't decide on which option is better and
just provide another setting (an advanced one in this case) and let the
user decide.