Make parsed link list local
There's no need to keep the list of parsed links as a global variable
(more specifically an instance variable of the global opts object). It's
main purpose is to initialize a list of coub objects in main(). The
links themselves aren't used afterwards.
Only the length of the list was used for progress messages (as the coub
list gets decimated upon retries), which can be easily replaced by using
a total count.