Introduce argparse
I'm starting to ask myself, why I'm even doing this. I do not like
argparse. It's clunky for anything but basic "assign value x to variable
 y". Unfortunately manual parsing doesn't scale well. An endless
 if-elif-else statement is not exactly easily readable and the more
 options you have the more additional tests are necessary to ensure
 valid input and arguments. Not to mention that some mutually exclusive
 options (e.g. --bestvideo and --worstvideo) wouldn't throw an error,
 but rather overwrite each other as they set the same variable.

So argparse it is despite its flaws.

The last few commits were done in preparation of the introduction of
argparse, but there were still a few details, which couldn't be
implemented easily with argparse and required additional changes.
Because of that this commit turned out larger than I initially expected
or wanted.

Here the list of additional changed:

  -) move default sort method to <container class>.__init__()

Previously the opts object was initialized before any command-line
parsing happened, so we could access its default_sort list during the
initialization of container objects. This list isn't globally available
anymore. Each container class is now responsible to set its default sort
method if it isn't specified by the user.

  -) combine preview and preview_command

It was unnecessarily complicated to have both a bool to activate
previewing and a string to hold the actual command. Instead just use the
 command string and set it to None if unspecified.

  -) throw error when container is initialized with an URL

Previously no_url() was used to throw errors during assignment, when an
URL was used with an input option. This check is now part of the
BaseContainer initialization.

  -) remove type-specific code from BaseContainer

Each type of container has its own class and should handle its own
quirks (e.g. restricting the max. page count to 99 for tags).

  -) disable --hot option

The --hot option was always the weird one, because it didn't use an
argument. The sort method was therefore added to the end of the option,
which basically led to 8 possible --hot options. I'm sure this behavior
can be emulated with argparse (e.g. specifying all 8 options and then
using a custom action to initialize the HotSection object with the right
sort method) or I come up with a different solution (e.g. sort method as
argument), but that's a challenge for a different commit. So for now
just disable the --hot option entirely.

Again barely tested.