Reintroduce --hot option
After giving it some thought I decided to go a different route with the
--hot option and treat it as a shortcut for downloading the hot section
with the default sort method (shortcut being relative, because just
using 'hot' - as mangled input URL - has the same effect).
If a different sort method is wanted, then there are more than enough
alternatives to specify it.

  -) https://coub.com#hot_daily
  -) coub.com#hot_weekly
  -) hot#fresh
  -) rising

The input help was adjusted to reflect those changes.

This commit also alters the way HotSection objects are initialized. Now
they only require a sort method up initialization (as opposed to a
identifier+sort string with the other containers).