Add support for featured community
I'm probably repeating myself, but Coub is sometimes just a mess. The
featured community is another good example for that.

The first problem is that there are 2 basic URLs to access the featured
section.

  1. https://coub.com/community/featured
  2. https.//coub.com/featured/coubs

The former is what Coub links in the side menu, so I deem it the
official URL. The latter provides the exact same coubs (also utilizes
the same API endpoint), but uses different URLs for each sort method.
So a lot of possibilities to consider in normalize_link().

The second problem is that the featured community is a community only by
name and URL template. Different API endpoint and different supported
sort methods.
So a lot of exceptions to make in Community.template().

Both of these functions grow considerably with this commit and it won't
get better, when I add support for the Coub of the Day community, which
is also comes with its own set of rules.