Re-introduce container duplicate checking
It fell prey to the transition process from CoubInputData to a more
modular container system. Now it is back and once again weeds out
containers, which share the same type, id and sort method.
The only exception to this rule is the random category, which obviously
shouldn't count as a duplicate, even if all its identifiers are the
same.