Add basic GUI
This is supposed to make CoubDownloader more accessible for people, who
are not comfortable using the command line. I can't stress enough though
that it is only very basic. CoubDownloader is first and foremost a CLI
tool. Any graphical frontend will always be an addition and not the main
focus.
I also make it a rule not to change the main script to accommodate to
the GUI. If an option doesn't make sense with a GUI, it's coub-gui.py's
job to either find a way around it or make sure it doesn't interfere
with the execution.

The README.md outlines the small differences regarding functionality, so
I refrain from repeating them here.

I'm also aware that this is only a small step in the right direction.
People still have to go through the trouble of installing Python and
some 3rd party packages to fully utilize CoubDownloader.