Add support for config files
To change the default options one had to edit the script itself. All
default settings are right at the beginning of the script accompanied by
short descriptions, so it's by no means difficult to change them. There
is one problem though, that interferes with some of my future plans.

What if one doesn't run CoubDownloader as a script via the Python
interpreter, but as a standalone executable (e.g. after bundling all
dependencies with PyInstaller)? In that case it would be impossible to
change the default options.
So let's add the ability to change default options with a config file as
well. The config file must be named "coub.conf" and be in the same
location as the main script (although more paths could be easily added).

Options are defined the same way as in DefaultOptions. White spaces are
stripped from both the option name and value and lines starting with "#"
are ignored as comments. Unknown option names or clearly invalid values
(e.g. not an int value for RETRIES) are ignored. So are config files,
which can't be read. Any errors will print warnings, but won't force a
hard exit.
However, it is important to note that (just like previously) there are
no thorough validity checks as far as default options are concerned. All
kinds of weird errors can occur, if one isn't careful to choose valid
values.