Change default output template from None to %id%
For a while now parse_cli() makes sure that %id% is converted to the
internal value None, which leads to faster skipping of existing files
if no archive file is used.
So we can change the default value to %id% without missing the increased
skipping speed, which makes more sense from a user's perspective and
allows us to fetch the actual default value for the help text.