Combine recoub-related options
Instead of using 2 different booleans, just use a value range to define
how much to prefer recoubs (similar to what is already used for AAC
audio selection).
The main motivation is that it removes a possible source for errors when
using custom default options, but it also simplifies one of the
workarounds in coub-gui.py.