Add validity test for default options
Until now I followed a simple logic when it came to default options. If
someone would take it upon themselves to change the default values in
the script, then this person has the sole responsibility to ensure that
the new values make sense.

This stance becomes somewhat difficult to justify, now that reading
custom default options (via coub.conf) is a central part of the script.
Measures must be taken to prevent invalid default values. This is the
goal of the new DefaultOptions.check_values() method.