Obfuscate default options in script
Using a config file should be the preferred way to change defaults and
example.conf describes all options in detail (more so than the comments
in the main script). That's why I'll stop treating the DefaultOptions
class special. It's moved further down and grouped with other class
definitions and the extensive descriptions were removed.
Of course it's still easy to change default options within the script,
but making it as easy as possible is not a priority anymore.