Check against archive during the parsing process
For normal users the process of checking coub IDs against an archive
file was efficient enough. But let's assume an "enthusiast" scrapes
communities on a regular basis. In that case they're not only likely to
work with several thousand input IDs per script invocation, but also to
have an archive file with an even greater magnitude.

To put it into numbers:
Skipping 99% of 90.000 IDs based on an archive file with 400.000 IDs
takes approximately 160 seconds.

That's not acceptable. Now we can do 2 things to make this faster.

  1.) Use a set to store the archive's content
  2.) Check the parsed ID list against the archive

1. is done for obvious reasons. 2. prevents the unnecessary creation of
Coub objects, all the existence checks during the processing chain and
the printing of an "exists" messages.

Going back to the previous example with the new approach:
Skipping 99% of 90.000 IDs based on an archive file with 400.000 IDs
takes approximately 0,8 seconds.