Add config support for GUI version
Some small adjustments had to be made to DefaultOptions' initialization
in the process.
When importing coub.py inside a coub-gui standalone binary, __file__
points to the PyInstaller extraction location and not the location of
the binary. So the binary location must be determined inside coub-gui.py
and then passed to the __init__ function.

This commit also makes sure that any custom FFmpeg path is converted to
an absolute path (wasn't yet the case for coub.py).