Replace forbidden characters in filename
This was already done on a very basic level, but now it replaces the
forbidden characters (instead of just removing them) and it covers a
slightly broader collection of forbidden characters, including the most
common pitfalls for filenames on Windows.
The latter is especially important because %creation% always led to the
fallback filename as it includes colons.

Still not perfect, but it should prevent many titles from failing.