Add special keywords for tag_sep and fallback_char
In particular it wasn't possible to use a whitespace as tag separator or
fallback character or to only remove disallowed characters from the
filename, if specified via a config file.
This is remedied by introducing keywords that the script will interpret
accordingly

  space -> whitespace (tag separator and fallback character)
  None  -> empty string (i.e. removal) (fallback character)

Not the most elegant solution, but we're only dealing with 2 keywords
anyway.