Fix confusing checkbox for --keep (but not really)
When setting "KEEP = True" via a config file, a checked box actually
meant False (don't keep), while an unchecked box meant True (keep). This
fixes the confusing selection, but then again not really, because the
option now suffers the same problem as with the CLI version, that it can
only be used to activate keeping and not deactivate it.

So to summarize:

  KEEP = False: Checked -> True, Unchecked -> False
  KEEP = True:  Checked -> True, Unchecked -> True

The only real solution would be to introduce a --no-keep option
(similar to --preview/--no-preview) for the CLI version and write a
proper GUI for the GUI version. The latter won't be happening any time
soon and I don't see the point of introducing yet another CLI option
that nobody will use (including myself).