Add retries for link container processing
Trying to reestablish a lost connection during the download phase has
been a feature for quite a while now. The parsing process of link
containers, however, was still prone to errors if the connection was
lost or terminated by Coub.
This commit adds a recursive call (very similar to the one already being
used in attempt_process()) to allow as many retries for a single
container as specified via --retries (so by default 5). That increases
the complexity of the process function a bit, so I might have to tidy it
up in the future, but for now it works and should make the script very
robust against connection issues.