Postpone exit caused by invalid defaults
Doing a hard exit as soon as an invalid default value was detected had
one major drawback for the GUI version: The user wouldn't see the error
unless the program was started from the command line (which most users,
who are interested in a GUI, won't do). In their eyes trying to execute
the GUI would just do nothing.

This can be remedied by exiting AFTER Gooey created the main window
(i.e. after GooeyParser.parse_args() was called). Then the error will be
printed in the output screen, just like config-related warnings before.
Unfortunately not all errors can be caught this way, since some invalid
defaults can make GooeyParser itself fail. With the new solution the
user will also see GooeyParser's traceback in such a case, but the
initial error message (printed in red) still stands out, so I consider
it a good tradeoff for being able to open the GUI in most cases.

Additionally all config-related warnings are promoted to errors and will
lead to an early exit as well. I believe that it's more beneficial to
the user to prevent the execution if something is wrong with the config,
instead of just printing a warning. Chances are the user will stop the
script anyway to quickly fix the erroneous parts.