Change library search path based on script usage
As stated in the README my primary reason to not provide Linux builds is
my inability to create standalone binaries, which work across a wide
range of Linux distributions. The main issue was subprocess.run() trying
to use the packaged libraries instead of the system ones.

As it turns out the fix for this is quite simple and well documented. I
swear I read the particular page in the documentation at least 3 times,
but either I was wrestling some other issue at the time or I was just to
clueless to realize that it specifically talks about the problem I was
experiencing.
As a side note, because subprocess.check_call() lacks the ability to
accept custom environments, all instances were replaced by
subprocess.run() with check=True (which does the same thing).

Anyway, this means that the way for a standalone build of the CLI tool
for Linux x86_64 is clear (already successfully tested on Debian,
Ubuntu, Fedora and CentOS). 32-bit won't be possible because staticx
lacks support for it (and I don't want to mess with min glibc versions).
I also haven't tested the GUI version yet, which was plagued by all
sorts of little problems last time I tried. But those are problems for
another day.